backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<3.13
psycopg2-binary>=2.9,<3.0  # For PostgreSQL, if used
python-dotenv>=0.19.0  # For loading environment variables
pandas>=1.3.0  # For data processing
numpy>=1.20.0  # For numerical operations
PyJWT>=2.3.0  # For JWT authentication
django-cors-headers>=3.10.0,<4.0  # For CORS support